Resolve WIT-4229 "Ability to execute a job with a different and configurable user"
# New features and improvements

**Made the “runAs” property of jobs configurable**
You can now specify the service principal name used to run jobs by setting the `specific.runAsPrincipalName` field in the provisioningRequest.

# Bug fixes

Fixed a bug in the output port reverse provisioning, where it was possible to encounter the error "Not able to convert data type 'LONG' in Open Metadata" during the reverse provisioning process.


# Related issue

Closes WIT-4229